+++ textblob_de/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
-from textblob_de.blob import TextBlobDE, Word, WordList, Sentence
+from textblob_de.blob import TextBlobDE, BlobberDE, Word, WordList, Sentence
 from textblob_de.tokenizers import NLTKPunktTokenizer, PatternTokenizer
 from textblob_de.taggers import PatternTagger
 from textblob_de.parsers import PatternParser
@@ -8,6 +8,6 @@ from textblob_de.np_extractors import PatternParserNPExtractor
 from textblob_de.sentiments import PatternAnalyzer
 
 
-__version__ = '0.2.2a2'
+__version__ = '0.2.2a3'
 __author__ = 'Markus Killer'
 __license__ = "MIT"